new_app/backend/requirements.txt
fastapi
uvicorn
pydantic
openai
python-dotenv
psycopg2-binary
marked
asyncio
email-validator
